requirements.txt
asgiref==3.5.2
chardet==5.0.0
colorama==0.4.5
Django==4.0.6
django-excel==0.0.10
django-jazzmin==2.6.0
django-modeltranslation==0.18.4
django-recaptcha==3.0.0
django-registration-redux==2.11
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
et-xmlfile==1.1.0
lml==0.1.0
openpyxl==3.0.10
Pillow==9.2.0
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pyexcel-xlsx==0.6.0
pypng==0.20220715.0
python-gettext==4.1
pytz==2022.2.1
qrcode==7.4.2
sqlparse==0.4.2
texttable==1.6.5
typing_extensions==4.7.1
tzdata==2022.1
whitenoise==6.2.0
